ByAzen 1.3.1, волна 1 идей + музыка 2.0 (ICY-эфир, режимы очереди, нормализация, автопауза, таймер сна, подборки, M3U, визуализатор, клавиши плеера)

+++ src/main/java/rtx/byazen/api/modules/impl/Utils/DeathHistoryModule.java
@@ -71,7 +71,7 @@ extends Module {
         double z = this.mc.player.getZ();
         String world = this.mc.world.getRegistryKey().getValue().getPath();
         String coords = Math.round(x) + " " + Math.round(y) + " " + Math.round(z);
-        DeathHistoryModule.Entry entry = new DeathHistoryModule.Entry(Math.round(x), Math.round(y), Math.round(z), world, System.currentTimeMillis());
+        DeathHistoryModule.Entry entry = new DeathHistoryModule.Entry((int) Math.round(x), (int) Math.round(y), (int) Math.round(z), world, System.currentTimeMillis());
         this.entries.add(0, entry);
         while (this.entries.size() > LIMIT) {
             this.entries.remove(this.entries.size() - 1);
